src/requirements.txt
boto3==1.34.89
requests==2.31.0
Pillow==10.3.0
